### 2020-06-29 #2406 @whedon
[PRE REVIEW]: starfish: scalable pipelines for image-based transcriptomics
**Submitting author:** @neuromusic (<a href="http://orcid.org/0000-0002-4638-7015">Justin Kiggins</a>)
**Repository:** <a href="https://github.com/spacetx/starfish" target ="_blank">https://github.com/spacetx/starfish</a>
**Version:** v0.2.1
