### 2020-06-29 #2406 @danielskatz
👋 @neuromusic

Thanks for your submission to JOSS. As described in our [blog post](https://blog.joss.theoj.org/2020/05/reopening-joss) announcing the reopening of JOSS, we're currently working in a "reduced service mode", limiting the number of papers assigned to any individual editor.

Since reopening JOSS, we've had a very large number of papers submitted and as such, yours has been put in our backlog that we have been and will continue to work through over the coming weeks and months.

Thanks in advance for your patience!
