+++ docs/source/help_and_reference/index.rst
@@ -13,6 +13,6 @@ Help & Reference
     working_with_starfish_outputs/index
     processing_at_scale/index
     configuration/index
-    glossary/glossarys
+    glossary/glossary
     spacetx_consortium/index
     request_support/index
