+++ README.rst
@@ -36,6 +36,9 @@ It lets you build scalable pipelines that localize and quantify RNA transcripts 
 generated by any FISH method, from simple RNA single-molecule FISH to combinatorial barcoded
 assays.
 
+- `Starfish enterprise: finding RNA patterns in single cells <https://www.nature.com/articles/d41586-019-02477-9>`_, Nature Technology Feature, Aug 2019
+- `Developing a Computational Pipeline and Benchmark Datasets for Image-Based Transcriptomics https://www.ascb.org/science-news/developing-a-computational-pipeline-and-benchmark-datasets-for-image-based-transcriptomics/`_ ASCB Science News, Dec 2018
+
 Documentation
 -------------
 
