+++ README.rst
@@ -37,16 +37,12 @@ generated by any FISH method, from simple RNA single-molecule FISH to combinator
 assays.
 
 - `Starfish enterprise: finding RNA patterns in single cells <https://www.nature.com/articles/d41586-019-02477-9>`_, Nature Technology Feature, Aug 2019
-- `Developing a Computational Pipeline and Benchmark Datasets for Image-Based Transcriptomics https://www.ascb.org/science-news/developing-a-computational-pipeline-and-benchmark-datasets-for-image-based-transcriptomics/`_ ASCB Science News, Dec 2018
+- `Developing a Computational Pipeline and Benchmark Datasets for Image-Based Transcriptomics <https://www.ascb.org/science-news/developing-a-computational-pipeline-and-benchmark-datasets-for-image-based-transcriptomics/>`_, ASCB Science News, Dec 2018
 
 Documentation
 -------------
 
-For the `quickstart <https://spacetx-starfish.readthedocs.io/en/latest/gallery/quick_start/plot_quick_start.html>`_,
-        `user guide <https://spacetx-starfish.readthedocs.io/en/latest/user_guide/>`_,
-        `examples <https://spacetx-starfish.readthedocs.io/en/latest/gallery/>`_,
-        and `API <https://spacetx-starfish.readthedocs.io/en/latest/api/>`_,
-        see `spacetx-starfish.readthedocs.io <https://spacetx-starfish.readthedocs.io/en/latest/>`_.
+For the `quickstart <https://spacetx-starfish.readthedocs.io/en/latest/gallery/quick_start/plot_quick_start.html>`_, `user guide <https://spacetx-starfish.readthedocs.io/en/latest/user_guide/>`_, `examples <https://spacetx-starfish.readthedocs.io/en/latest/gallery/>`_, and `API <https://spacetx-starfish.readthedocs.io/en/latest/api/>`_, see `spacetx-starfish.readthedocs.io <https://spacetx-starfish.readthedocs.io/en/latest/>`_.
 
 Installation
 ------------
@@ -62,9 +58,9 @@ starfish supports python 3.6 and above (python 3.7 recommended) and can easily b
 Contributing
 ------------
 
-We welcome contributions from our users! See our contributing.rst_ and documentation_ for more information.
+We welcome contributions from our users! See our contributing.rst_ and `developer guide`_ for more information.
 
-.. _documentation: https://spacetx-starfish.readthedocs.io/en/latest/
+.. _`developer guide`: https://spacetx-starfish.readthedocs.io/en/latest/developer_guide/
 .. _contributing.rst: https://github.com/spacetx/starfish/blob/master/CONTRIBUTING.rst
 
 Help, support, and questions
