### 2020-06-30 78ab15d
cleaned up user guide of To Dos and vestigial links. added some examples to removing autofluorescence and illumination correction.
