### 2020-06-30 e0eb180
cleaned up user guide
* removed low priority unfinished examples and sections
* removed old redundant content
* added white top-hat examples
