+++ examples/tutorials/plot_image_corrections.py
@@ -1,196 +1,0 @@
-"""
-Image Corrections
-=================
-.. container:: toggle
-
-    .. container:: header
-
-        **Show/Hide Code**
-
-    .. code-block:: xml
-       :linenos:
-
-       import matplotlib.pyplot as plt
-       import numpy as np
-       import xarray as xr
-
-       import starfish
-       import starfish.data
-       from starfish.types import Axes
-
-       experiment = starfish.data.ISS(use_test_data=True)
-       image: starfish.ImageStack = experiment['fov_001'].get_image('primary')
-
-       image_2d = image.sel({Axes.CH: 0, Axes.ROUND: 0, Axes.ZPLANE: 0})
-
-       plt.imshow(np.squeeze(image_2d.xarray.values))
-       plt.show()
-"""
-
-
-###################################################################################################
-# .. _tutorial_illumination_correction:
-#
-# Illumination Correction
-# =======================
-#
-# The goal of illumination correction is to remove uneven illumination of the image caused by non
-# uniform illumination of the field of view, characteristics of the sensor, (like vignetting), or
-# orientation of the tissue’s surface with respect to the light source.
-#
-# The simplest forms of illumination correction are called “prospective correction” and are based on
-# background subtraction. This involves taking additional images using the microscopy apparatus to
-# help calibrate. These can either be acquired by averaging a series of images captured with no sample
-# and no light (dark image), or with no sample and light (bright image).
-#
-# Starfish can apply this type of background correction by exposing the :py:class:`ElementWiseMult`
-# :ref:`Filter <filtering>`. The user is responsible for transforming their calibration images into
-# the correct matrix to correct for background, and then :py:class:`ElementWiseMult` can apply a
-# transformation to correct any uneven illumination.
-#
-# The below plot shows a single plane of an in-situ sequencing experiment.
-#
-# .. container:: toggle
-#
-#    .. container:: header
-#
-#       **Show/Hide Code**
-
-import matplotlib.pyplot as plt
-import numpy as np
-import xarray as xr
-
-import starfish
-import starfish.data
-from starfish.types import Axes
-
-experiment = starfish.data.ISS(use_test_data=True)
-image: starfish.ImageStack = experiment['fov_001'].get_image('primary')
-
-image_2d = image.sel({Axes.CH: 0, Axes.ROUND: 0, Axes.ZPLANE: 0})
-
-plt.imshow(np.squeeze(image_2d.xarray.values))
-plt.show()
-
-###################################################################################################
-# This image was corrected before it was sent to us, but we can introduce an uneven illumination
-# profile. Below we mock an extremely severe 200% decrease in illumination from left to right.
-
-lightness = np.linspace(4, 1, image_2d.xarray.sizes[Axes.X])
-gradient_data = np.tile(lightness, reps=(image_2d.xarray.sizes[Axes.Y], 1))
-gradient = xr.DataArray(
-    data=gradient_data[np.newaxis, np.newaxis, np.newaxis, :, :],
-    dims=(Axes.ROUND.value, Axes.CH.value, Axes.ZPLANE.value, Axes.Y.value, Axes.X.value)
-)
-
-# introduce the gradient, overwriting the ImageStack
-data = image_2d.xarray.values / gradient.values
-image_2d = starfish.ImageStack.from_numpy(data)
-
-# display the resulting image
-plt.imshow(np.squeeze(image_2d.xarray.values))
-plt.show()
-
-###################################################################################################
-# The illumination profile has increased the intensity of the background in the right side of the
-# image. This is problematic for many spot finding methods that set thresholds for peak intensities
-# globally across the image; spots can be incorrectly excluded in low-illumination areas, and this
[…]
