### 2020-07-05 #2440 @whedon
[REVIEW]: starfish: scalable pipelines for image-based transcriptomics
**Submitting author:** @neuromusic (<a href="http://orcid.org/0000-0002-4638-7015">Justin Kiggins</a>)
**Repository:** <a href="https://github.com/spacetx/starfish" target ="_blank">https://github.com/spacetx/starfish</a>
**Version:** 0.2.2
