### 2020-07-15 spacetx/starfish#1896 @giovp
Tutorial "Showing Image of ImageStack Plane" does not run
#### Description
Running codeblock from tutorial [Showing Image of ImageStack Plane](https://spacetx-starfish.readthedocs.io/en/latest/gallery/how_to/imshow_plane.html#tutorial-imshow-plane) gets stucked. No error/warning neither if run on jupyter notebooks or python command line

#### Steps/Code to Reproduce
```python
import starfish.data
from starfish import FieldOfView
from starfish.types import Axes
experiment = starfish.data.STARmap(use_test_data=True)
stack = experiment.fov().get_image(FieldOfView.PRIMARY_IMAGES)
print(stack)
```
#### Expected Results
Something like:
```
<starfish.ImageStack (r: 4, c: 4, z: 1, y: 140, x: 200)>
```

#### Actual Results
Nothing

#### Versions
```
Darwin-18.7.0-x86_64-i386-64bit
Python 3.7.7 (default, May  6 2020, 04:59:01)
[Clang 4.0.1 (tags/RELEASE_401/final)]
NumPy 1.19.0
SciPy 1.5.1
scikit-image 0.15.0
pandas 1.0.5
sklearn 0.23.1
xarray 0.16.0
sympy 1.6.1
starfish 0.2.1
```

With `ISS` or `DARTFISH` it does load the fov image but throws another error afterward (which I suspect it's because the rest of the codeblock of the tutorial is not supposed to be run on such data

This issue references the JOSS submission openjournals/joss-reviews#2440
