### 2020-07-15 spacetx/starfish#1897 @giovp
Tutorials "How to normalize" improvements.
These points refers to the tutorial on [normalization strategies](https://spacetx-starfish.readthedocs.io/en/latest/user_guide/#how-to-normalize)
* Tutorials on normalization are really well made, but having the images alongside the text would greatly improve the readability.
* The tutorial on normalizing by matching histograms is really nice for clearly showing when *not* to do it, but at least one example where the histograms match, and so it would be appropriate to use that normalization strategy, would be useful (especially since, it seems to me, is really knowledge-based and analyst without much experience would not have that intuition).

**Describe the solution you'd like**
Images alongside text in the tutorial

Running the tutorials on a local jupyter instance it's definitely fine, and works well. Feel free to disregard this if the motivation for this choice is already explained somewhere else (and i missed it).

**Additional context**

This issue references the JOSS submission openjournals/joss-reviews#2440
