### 2020-07-15 spacetx/starfish#1896 @giovp
For same reason (cannot import STARmap dataset) I can't run codeblock in tutorial [comparison find posts algorithms](https://spacetx-starfish.readthedocs.io/en/latest/gallery/tutorials/spotbased_decoding.html#comparison-of-findspotsalgorithms)
