### 2020-07-17 spacetx/starfish#1897 @neuromusic
Thank you for the suggestions, @giovp!

This is due to the resource constraints of ReadTheDocs rendering the tutorials... We're tracking this (and exploring alternatives) in #1865

> at least one example where the histograms match, and so it would be appropriate to use that normalization strategy, would be useful

That's a good suggestion. We'll see what we can do. cc @mattcai
