### 2020-07-17 spacetx/starfish#1896 @neuromusic
This example dataset is rather large (`<slicedimage.TileSet (r: 6, c: 4, z: 28, x: 1024, y: 1024)>`) & I suspect that it's just taking time to download:

On my connection, it took 16 min 30 sec.

``` Python
In [10]: %timeit -n1 -r1 stack = experiment.fov().get_image(FieldOfView.PRIMARY_IMAGES)
16min 30s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
```

Note: Remote files are cached in `~/.starfish/cache`, so the second time this is run (on the comparisons tutorial), it will load more quickly (about 7 seconds on my machine)
