### 2020-07-17 spacetx/starfish#1898 @giovp
Tutorial "Pre-defined Watershed Segmentation Pipeline"
The text on the tutorial [Pre-defined Watershed Segmentation Pipeline](https://spacetx-starfish.readthedocs.io/en/latest/gallery/tutorials/watershed_segmentation.html#pre-defined-watershed-segmentation-pipeline) seems cut. Explanation of how the two modules differ, first paragraph, not very clear what it meant to say

Checked on both stable and latest build, not sure whether that's a bug

**Describe the solution you'd like**
rest of the sentence

**Additional context**
This issue references the JOSS submission openjournals/joss-reviews#2440
