### 2020-07-17 spacetx/starfish#1897 @giovp
Great, thank you! Did not notice the open issue. Nevermind, I think just running tutorials locally is also totally fine.

Regarding the example, I think it would improve readability and clarity but really marginal. what is there is already great, thank you!
