### 2020-07-17 spacetx/starfish#1899 @giovp
Tutorial "Loading ImageJ ROI Sets"
In the tutorial [Loading ImageJ ROI Sets](https://spacetx-starfish.readthedocs.io/en/latest/gallery/tutorials/manual_segmentation.html#tutorial-manual-segmentation) it's unclear how to get/generate the ROI.zip file.

The explanation for ImageJ seems quite clear, but it's not clear what image to use for that specific task. I also looked for the image in `~/.starfish/cache` but could not find any pointer to the MERFISH data (which I assume is the one needed for the tutorial). This is also related to the [assigning spot to cells tutorial](https://spacetx-starfish.readthedocs.io/en/latest/gallery/tutorials/assigning_spots.html#assigning-spots-to-cells)

**Describe the solution you'd like**
either include ROI in `starfish.data` or better explanation how to generate the files.

**Additional context**
This issue references the JOSS submission openjournals/joss-reviews#2440
