### 2020-07-19 spacetx/starfish#1900 @vals
More intuitive ordering for the link to documentation
I got a bit lost finding the documentation. I suggest changing

> For the quickstart, user guide, examples, and API, see spacetx-starfish.readthedocs.io.

to

> See spacetx-starfish.readthedocs.io for the quickstart, user guide, examples, and API

Related to https://github.com/openjournals/joss-reviews/issues/2440
