### 2020-07-19 spacetx/starfish#1901 @vals
Quickstart and napari
The 'default' installation instruction on the Readme for the repository simply says to install with `$ pip install starfish`. But the quickstart guide requires Starfish to have been installed with `$ pip install starfish[napari]`.

It would be better either if Napari wasn't required for the quick start, or included with the default installation.

Related to https://github.com/openjournals/joss-reviews/issues/2440
