### 2020-07-20 spacetx/starfish#1902 @vals
Update open napari window?
From going through the quickstart, it seems a new napari window is opened each time you call `display()`? Is there a way to have it update the already open window instead?

Related to https://github.com/openjournals/joss-reviews/issues/2440
