### 2020-07-20 spacetx/starfish#1903 @vals
Displaying example from data formatting tutorial
After going through the tutorial on how to format image data for Starfish (https://spacetx-starfish.readthedocs.io/en/latest/gallery/data_formatting/plot_format_structured_data.html#sphx-glr-gallery-data-formatting-plot-format-structured-data-py) I wanted to use the `display()` function to look at it. But if you do e.g.
```
imgs = e.fov().get_image('primary')
display(nuclei)
```
napari launches with 'Not Responding' status and crashes the Python interpreter when you kill the window.

Related to https://github.com/openjournals/joss-reviews/issues/2440
