### 2020-07-20 spacetx/starfish#1904 @vals
Include output in example notebooks
The notebooks with examples of using Starfish for different experiments at https://github.com/spacetx/starfish/tree/master/notebooks seem nice. From those it seems pretty straightforward to adapt settings etc to a new experiment. But it would be a better resource if the output from running the notebooks would be included. This way one can learn more with different functions do when e.g. scaling data without having to download and run the notebook yourself.

Related to https://github.com/openjournals/joss-reviews/issues/2440
