### 2020-07-20 spacetx/starfish#1905 @vals
JOSS paper does not mention alternative tools
The JOSS paper should mention how this tool compares to other tools.

As far as I know there are no other tools that do what starfish does. But it would be good to state this in the paper.

Related to https://github.com/openjournals/joss-reviews/issues/2440
