### 2020-07-21 #2440 @vals
First round finished from my end as well. I would like all my issues addressed. But most importantly from my end there should be versions of the practical examples that can be read without running the notebooks ( https://github.com/spacetx/starfish/issues/1904 ). Even when napari is used, the authors could make versions with screenshots. There is a lot of documentation about all the functionality, but without examples it is hard to get an understanding what the different functionality does.
