### 2020-07-21 spacetx/starfish#1899 @neuromusic
Thanks for calling this out, @giovp !!

We want to explicitly demonstrate loading from a file, so here's the plan:
- Update the example to download `RoiSet.zip` from the [blob on Github](https://github.com/spacetx/starfish/blob/b3aa7758f3d60c2c28083cfa1cf72e954d294592/starfish/core/morphology/binary_mask/test/RoiSet.zip)
- load from the downloaded file
