### 2020-07-23 0759fe7
updated sympy version in requirements.txt because newest sympy==1.6.* has a threading bug that breaks LocalMaxPeakFinder.
