+++ REQUIREMENTS.txt
@@ -21,7 +21,7 @@ scikit-learn
 scipy
 showit >= 1.1.4
 slicedimage==4.1.1
-sympy
+sympy ~= 1.5.0
 tqdm
 trackpy
 validators
