### 2020-07-23 04a4989
addresses JOSS reviwer's comments in issue #1898
