### 2020-07-23 84f9385
addresses JOSS reviwer's comments in issue #1900
