### 2020-07-23 spacetx/starfish#1901 @mattcai
Since many users find napari to be very valuable, we want to keep napari in the quickstart guide.  I will update the default installation instructions instead.
