### 2020-07-23 fa8772d
addresses JOSS reviwer's comments in issue #1901
