### 2020-07-23 725df70
addresses JOSS reviwer's comments in issue #1902
