### 2020-07-23 e42e071
Fix localmaxpeakfinder spot_props filter (#1839)
* modified local_max_peak_finder to be able to remove labeled regions from masked_image when masked_image is 2D

* fixed to allow 2D data with is_volume=True

* added note LocalMaxPeakFinder on 3D data with is_volume=True is slow

* removed unnecessary code

* Update local_max_peak_finder.py

* Delete expression
