### 2020-07-24 edb5350
Address easy fixes by JOSS reviewers (#1906)
* updated sympy version in requirements.txt because newest sympy==1.6.* has a threading bug that breaks LocalMaxPeakFinder.

* addresses JOSS reviwer's comments in issue #1898

* addresses JOSS reviwer's comments in issue #1900

* addresses JOSS reviwer's comments in issue #1901

* addresses JOSS reviwer's comments in issue #1902
