### 2020-07-24 spacetx/starfish#1907 @shazanfar
JOSS MS comments
Related to https://github.com/openjournals/joss-reviews/issues/2440

Thanks for writing this software, I think the starfish & its associated documentation represents a huge amount of work and the authors as well as all involved should be commended for their efforts!

MS:

Since other data modalities are also measured (e.g. protein), it's worth using more general wording than "gene expression", perhaps "abundance" is more relevant.

It's difficult to assess from the JOSS MS and the software documentation who exactly is the target audience. While indeed anyone with Python + some basic knowledge + starfish installed can run the Quick Start, I think the learning curve becomes quite steep once you start going through other examples and tutorials. It seems to me that either starfish is more suited to practitioners with at least some image analysis experience (or knowledge of the concepts behind), or starfish requires some more automated quality metrics/troubleshooting to be available & described.

Documentation/MS:

It's quite nice how various technologies' data formats are listed and examples shown + used. This being said, surely a number of technologies input formats are still under active development, and likely to change in future. How do you ensure that various file format updates for technologies are dealt with? Is it within scope of starfish developers to provide scripts for different versions of input files? I think answers to these should be mentioned in the Tutorials and Examples area.

I don't think seaborn is listed as a dependency, but is called at the end of the Quick start tutorial, it's worth either including as a dependency or mentioning it's needed at the beginning of the Quick start tutorial.

It's unclear from the User Guide how one could extract all relevant information into an external file format, to be used in other environments (e.g. R/Bioconductor). This is actually given in "Working with Starfish outputs" under "Help & Reference", but IMO could appear in the Tutorials and Examples area for greater visibility.

I can't see clear examples in the User Guide + Tutorials showcasing the scalability of starfish, I'd suggest adding a tutorial/how-to for starfish in such a context. Again, it's worth just making the "Help & Reference" section more visible, esp the 'Processing with AWS' section.

Cheers,
Shila
