### 2020-07-24 #2440 @shazanfar
Thanks for your patience @csoneson ! I've now completed the first round. I agree with @giovp and @vals comments regarding example data and including output in the tutorials. I've also requested amendments & clarification to do with scope + target audience, which I think would make the documentation more easily accessible for users. Thanks again! Cheers, Shila
