### 2020-07-28 spacetx/starfish#1903 @mattcai
I am not able to recreate your issue. Does the `display()` function work for you when you run the quickstart tutorial?

After running the code in [format structred data tutorial](https://spacetx-starfish.readthedocs.io/en/latest/gallery/data_formatting/plot_format_structured_data.html#sphx-glr-gallery-data-formatting-plot-format-structured-data-py) I am able to look at the (blank) images with:

```
from starfish import display
imgs = exp.fov().get_image('primary')
viewer = display(imgs)
```
