### 2020-07-28 spacetx/starfish#1904 @mattcai
This will be addressed by merging #1892 and solving the readthedocs resource limitations for rendering gallery examples (#1865).
