### 2020-07-28 #2440 @neuromusic
Quick question: How would you prefer we handle closing issues when we feel like we've addressed them? Should we close them or is it best to wait for our reviewers to close them when they are happy?
