### 2020-07-29 #2440 @csoneson
@neuromusic - if you are confident that the issue has been addressed I think you can close the issue, with a comment explaining the changes that were made. You can also consult with the reviewer first, in case additional clarifications are needed.
