### 2020-07-29 e2aafa1
added how-to for downloading RoiSet.zip to manual_segmentation.py and assigning_spots.py examples.
