### 2020-07-29 spacetx/starfish#1899 @mattcai
In order to preserve the ability for readthedocs to render example gallery in the future (#1865), we chose to comment out the code that downloads `RoiSet.zip` from github.
