### 2020-07-30 0e3c332
added current state of rnascope_hiplex.py pipeline in gallery along with feedback from ACDBio.
