+++ examples/pipelines/rnascope_hiplex.py
@@ -2,138 +2,562 @@
 .. _RNAScope_processing_example:
 
 Single Field of View for RNAScope HiPlex
-=========================================
+========================================
 
-This notebook walks through a work flow that analyzes one field of view of _____ from
-ACDBio, using the starfish package.
+ACDBio has a 12-gene HiPlex assay that uses 3 rounds of fluorescence imaging with 4 fluorescence RNA
+channels and 1 DAPI-stained nuclei channel. In the example data provided by ACDBio, these 15 images
+are saved as single-plane RGB tiffs with filenames that indicate the round and channel. A key for
+mapping images to genes is included in list.txt. This vignette will demonstrate how to format the
+data into SpaceTx format, register the images into the same coordinate system, and process them
+into cell x gene expression matrices using modules contained in starfish.
 
-This example processes an experiment with a single round from a single field of view of RNAScope
-HiPlex data taken from ____. These data
+Feedback from ACDBio:
+* Blob detection needs improvement in:
+  * Low-contrast regions - skip background subtraction and enhance contrast instead
+  * Clustered regions - use pre-defined spot size to break up clusters and count as single spots
+  * High-intensity saturated regions - use pre-defined spot size and intensity to to resolve
+  saturated regions as single spots
+  * Airy disks - either remove airy disks prior to spot finding or modify spot finding algorithm
+  to account for airy disk pattern
+* Image registration could be improved with rotation (low priority)
+* Cell Segmentation (watershed)
+  * sum intensities instead of max
+  * connect nuclear fragments with morphological operation
+* Expression Matrix
+  * Summary of data like dot size, intensity, etc - this is in the DecodedIntensityTable object
+  * Further analysis e.g. normalization - not in the scope of starfish
 
-The data consist of 15 images from 3 round, 5 channels, and 1 z-plane. Each image is
-(2048, 2048) (y, x).
+* Show blob results overlaid on raw
+
+* overlay_spot_calls() wrong coordinates?
 """
 
-from typing import Optional, Tuple
-from IPython import get_ipython
+###################################################################################################
+# Convert RGB images to greyscale images
+# --------------------------------------
+# starfish and the SpaceTx format expect greyscale images so the RGB data must first be converted.
+# This can be done in many tools, such as MATLAB and ImageJ. Here we chose to use Python Imaging
+# Library (PIL):
 
-import starfish
-import starfish.data
-from starfish import FieldOfView, DecodedIntensityTable
-from starfish.types import TraceBuildingStrategies
+import os
+import numpy as np
+from shutil import copy2
+from PIL import Image
 
-# equivalent to %gui qt
-ipython = get_ipython()
-ipython.magic("gui qt5")
 
+def convert_rgb_to_greyscale(input_dir, output_dir):
+    """
+    Original example data is stored in a directory named 'original_data' and consists of 15 RGB
+    images and 1 text file.
+    Copy everything into a new `output_dir` and convert any RGB tiffs to greyscale.
+
+    └── original_data
+            ├── list.txt
+            ├── R1_DAPI.tif
+            ├── R1_Probe T1.tif
+            ├── R1_Probe T2.tif
+            ├── R1_Probe T3.tif
+            ├── R1_Probe T4.tif
+            ├── R2_DAPI.tif
+            ├── R2_Probe T5.tif
+            ├── R2_Probe T6.tif
+            ├── ...
+
+    """
+
+    abs_input_dir = os.path.abspath(input_dir)
+    abs_output_dir = os.path.abspath(output_dir)
[…]
