### 2020-08-25 bcdf35f
add example pipelines to sphinx gallery (#1892)
* added baristaseq pipeline

* removed starfish usage

* converted notebooks to examples in pipelines gallery. only seqfish pipeline omitted because it is still a work in progress.

* added example pipelines gallery to landing page
