+++ examples/pipelines/osmfish_pipeline.py
@@ -1,0 +1,179 @@
+"""
+.. _osmfish_example:
+
+Reproduce published osmFISH results with starfish
+=================================================
+
+osmFISH is an image based transcriptomics technique that can spatially resolve tens of RNA
+transcripts and their expression levels **in situ**. The protocol and data analysis are described in
+this `publication`_. This notebook walks through how to use starfish to process the raw images
+from an osmFISH experiment into a spatially resolved gene expression image. We verify that
+starfish can accurately reproduce the results from the authors' original Python `pipeline`_
+
+.. _publication: https://www.nature.com/articles/s41592-018-0175-z
+.. _pipeline: http://linnarssonlab.org/osmFISH/image_analysis/
+"""
+
+from IPython import get_ipython
+import matplotlib
+import matplotlib.pyplot as plt
+
+# equivalent to %gui qt and %matplotlib inline
+ipython = get_ipython()
+ipython.magic("gui qt5")
+ipython.magic("matplotlib inline")
+
+matplotlib.rcParams["figure.dpi"] = 150
+
+###################################################################################################
+# Load Data into starfish from the cloud
+# --------------------------------------
+# The data from an osmFISH experiment are similar in form to a standard smFISH experiment. For each
+# round, each color channel corresponds to presence of a particular gene. Across rounds, the color
+# channels index different genes. Here, we analyze one FOV from the first round (r) and first
+# channel (c), which consists of 45 z-planes (z). Each image in this image stack is of dimensions
+# 2048x2048. The data are taken from mouse somatosensory cortex, and the gene in this channel is
+# Adloc.
+
+from starfish import data
+from starfish import FieldOfView
+
+experiment = data.osmFISH(use_test_data=True)
+imgs = experiment["fov_000"].get_image(FieldOfView.PRIMARY_IMAGES)
+print(imgs)
+
+###################################################################################################
+# Filter and visualize data
+# -------------------------
+# First, we remove background signal using a gaussian high-pass filter.
+
+from starfish.image import Filter
+
+filter_ghp = Filter.GaussianHighPass(sigma=(1, 8, 8), is_volume=True)
+imgs_ghp = filter_ghp.run(imgs, in_place=False)
+
+###################################################################################################
+# Next, we enhance the spots by filtering with a Laplace filter.
+
+filter_laplace = Filter.Laplace(sigma=(0.2, 0.5, 0.5), is_volume=True)
+imgs_ghp_laplace = filter_laplace.run(imgs_ghp, in_place=False)
+
+###################################################################################################
+# Finally, we take a maximum projection over z, which effectively mitigates effects of out of focus
+# z-planes.
+
+from starfish.types import Axes
+
+mp = imgs_ghp_laplace.reduce({Axes.ZPLANE}, func="max")
+
+###################################################################################################
+# We can now visualize our data before and after filtering.
+
+from numpy import np
+
+single_plane = imgs.reduce({Axes.ZPLANE}, func="max").xarray.sel({Axes.CH:0}).squeeze()
+single_plane_filtered = mp.xarray.sel({Axes.CH: 0}).squeeze()
+
+plt.figure(figsize=(10, 10))
+plt.subplot(121)
+plt.imshow(single_plane, cmap='gray', clim=list(np.percentile(single_plane.data, [1, 99.9])))
+plt.axis('off')
+plt.title('Original data, Round:0, Channel: 0')
+plt.subplot(122)
+plt.imshow(single_plane_filtered, cmap='gray', clim=list(np.percentile(single_plane_filtered.data, [1, 99.9])))
+plt.title('Filtered data, Round:0, Channel: 0')
+plt.axis('off')
+
[…]
