+++ examples/pipelines/merfish_pipeline.py
@@ -1,0 +1,233 @@
+"""
+.. _merfish_example:
+
+Reproduce published MERFISH results with starfish
+=================================================
+
+Multiplexed Error Robust Fish (MERFISH) is an image based transcriptomics technique that can
+spatially resolve hundreds to thousands of RNA species and their expression levels in-situ. The
+protocol and data analysis are described in this `publication`_. This notebook walks through how
+to use starfish to process the raw images from a MERFISH experiment into a spatially resolved
+cell by gene expression matrix. We verify that Starfish can accurately reproduce the results from
+the current Matlab based MERFISH `pipeline`_.
+
+.. _publication: https://science.sciencemag.org/content/348/6233/aaa6090
+.. _pipeline: https://github.com/ZhuangLab/MERFISH_analysis
+"""
+
+from IPython import get_ipython
+import matplotlib
+import matplotlib.pyplot as plt
+
+# equivalent to %gui qt and %matplotlib inline
+ipython = get_ipython()
+ipython.magic("gui qt5")
+ipython.magic("matplotlib inline")
+
+matplotlib.rcParams["figure.dpi"] = 150
+
+###################################################################################################
+# Load Data
+# ---------
+# The example data here correspond to DARTFISHv1 2017. The data represent human brain tissue from
+# the human occipital cortex from 1 field of view (FOV) of larger experiment. The data from one
+# field of view correspond to 18 images from 6 imaging rounds (r) 3 color channels (c) and 1 z-plane
+# (z). Each image is 988x988 (y,x)
+
+import pprint
+from starfish import data
+from starfish import FieldOfView
+
+experiment = data.MERFISH(use_test_data=False)
+
+pp = pprint.PrettyPrinter(indent=2)
+pp.pprint(experiment._src_doc)
+
+# note the structure of the 5D tensor containing the raw imaging data
+imgs = experiment.fov().get_image(FieldOfView.PRIMARY_IMAGES)
+print(imgs)
+
+###################################################################################################
+# Visualize codebook
+# ------------------
+# The MERFISH codebook maps each barcode to a gene (or blank) feature. The barcodes are 16 bit
+# vectors that can be read out, for each pixel, from the 8 rounds and 2 color channels. The codebook
+# contains a precise specification of how each of these 16 bit barcode vectors relate to the 5D
+# tensor of raw image data.
+
+experiment.codebook
+
+###################################################################################################
+# Visualize raw data
+# ------------------
+# A nice way to page through all this data is to use the display command. We have commented this out
+# for now, because it will not render in the docs. Instead, we simply show an image from the first
+# round and color channel.
+
+# display(imgs)
+
+from starfish.types import Axes
+
+single_plane = imgs.sel({Axes.ROUND: 0, Axes.CH: 0, Axes.ZPLANE: 0})
+single_plane = single_plane.xarray.squeeze()
+plt.figure(figsize=(7, 7))
+plt.imshow(single_plane, cmap='gray')
+plt.title('Round: 0, Channel: 0')
+plt.axis('off')
+
+###################################################################################################
+# Filter and scale raw data before decoding into spatially resolved gene expression
+# ---------------------------------------------------------------------------------
+# A high pass filter is used to remove background signal, which is typically of a low frequency.
+# This serves to remove autoflourescence, thus enhancing the ability to detect the RNA molecules.
+
+from starfish.image import Filter
+
+ghp = Filter.GaussianHighPass(sigma=3)
+high_passed = ghp.run(imgs, verbose=True, in_place=False)
+
[…]

[…]
