### 2020-08-25 spacetx/starfish#1903 @neuromusic
Hey @vals, can you confirm that this is an issue on the quickstart tutorial, too?

If you continue to see this bug, can you post the output of the following...

``` Python
import platform; print(platform.platform())
import sys; print("Python", sys.version)
import numpy; print("NumPy", numpy.__version__)
import scipy; print("SciPy", scipy.__version__)
import skimage; print("scikit-image", skimage.__version__)
import pandas; print("pandas", pandas.__version__)
import sklearn; print("sklearn", sklearn.__version__)
import xarray; print("xarray", xarray.__version__)
import sympy; print("sympy", sympy.__version__)
import starfish; print("starfish", starfish.__version__)
import napari; print("napari", napari.__version__)
```

thanks!
