+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,0 +1,7 @@
+# Ref: https://help.github.com/en/github/building-a-strong-community/configuring-issue-templates-for-your-repository#configuring-the-template-chooser
+blank_issues_enabled: true  # default
+contact_links:
+- name: 🤷 How do I?
+  url: https://forum.image.sc/tag/starfish
+  about: |
+    Please ask questions about using starfish on the image.sc forum.
