+++ README.rst
@@ -31,6 +31,8 @@ starfish: scalable pipelines for image-based transcriptomics
 
 -------------------------------
 
+SILLY CHANGE
+
 *starfish* is a Python library for processing images of image-based spatial transcriptomics.
 It lets you build scalable pipelines that localize and quantify RNA transcripts in image data
 generated by any FISH method, from simple RNA single-molecule FISH to combinatorial barcoded
