### 2020-11-05 #2440 @csoneson
👋 @neuromusic - just checking in that we're still on track to have these issues addressed within the next week.
