### 2020-11-10 #2440 @csoneson
> 👋 @neuromusic - just checking in that we're still on track to have these issues addressed within the next week.

ping @neuromusic - could you please give us an update on your progress?
