### 2020-11-13 spacetx/starfish#1904 @neuromusic
Thank you for the recommendation, @vals & @shazanfar. I agree that this would be highly valuable, however, we can't support this at this time. The datasets that these notebooks process are quite large and the RAM and time needed to render them far exceeds the resource limits of ReadTheDocs for open source projects. Details are captured in #1865.
