### 2020-11-13 spacetx/starfish#1903 @neuromusic
I'm going to close this issue, but please feel free to re-open @vals if you continue to see the problem.
