### 2020-11-13 0d97307
adds references to other tools. fixes #1905
