+++ paper/paper.bib
@@ -224,3 +224,34 @@
   school = {UC San Diego},
   URL = {https://escholarship.org/uc/item/5bq3128f},
 }
+
+@article{SMART-Q,
+    author = {Yang, Xiaoyu AND Bergenholtz, Seth AND Maliskova, Lenka AND Pebworth, Mark-Phillip AND Kriegstein, Arnold R. AND Li, Yun AND Shen, Yin},
+    journal = {PLOS ONE},
+    publisher = {Public Library of Science},
+    title = {SMART-Q: An Integrative Pipeline Quantifying Cell Type-Specific RNA Transcription},
+    year = {2020},
+    month = {04},
+    volume = {15},
+    url = {https://doi.org/10.1371/journal.pone.0228760},
+    pages = {1-11},
+    abstract = {Accurate RNA quantification at the single-cell level is critical for understanding the dynamics of gene expression and regulation across space and time. Single molecule FISH (smFISH), such as RNAscope, provides spatial and quantitative measurements of individual transcripts, therefore, can be used to explore differential gene expression among a heterogeneous cell population if combined with cell identify information. However, such analysis is not straightforward, and existing image analysis pipelines cannot integrate both RNA transcripts and cellular staining information to automatically output cell type-specific gene expression. We developed an efficient and customizable analysis method, Single-Molecule Automatic RNA Transcription Quantification (SMART-Q), to enable the analysis of gene transcripts in a cell type-specific manner. SMART-Q efficiently infers cell identity information from multiplexed immuno-staining and quantifies cell type-specific transcripts using a 3D Gaussian fitting algorithm. Furthermore, we have optimized SMART-Q for user experiences, such as flexible parameters specification, batch data outputs, and visualization of analysis results. SMART-Q meets the demands for efficient quantification of single-molecule RNA and can be widely used for cell type-specific RNA transcript analysis.},
+    number = {4},
+    doi = {10.1371/journal.pone.0228760}
+}
+
+@article{dotdotdot,
+    author = {Maynard, Kristen R and Tippani, Madhavi and Takahashi, Yoichiro and Phan, BaDoi N and Hyde, Thomas M and Jaffe, Andrew E and Martinowich, Keri},
+    title = "{dotdotdot: an automated approach to quantify multiplex single molecule fluorescent in situ hybridization (smFISH) images in complex tissues}",
+    journal = {Nucleic Acids Research},
+    volume = {48},
+    number = {11},
+    pages = {e66-e66},
+    year = {2020},
+    month = {05},
+    abstract = "{Multiplex single-molecule fluorescent in situ hybridization (smFISH) is a powerful method for validating RNA sequencing and emerging spatial transcriptomic data, but quantification remains a computational challenge. We present a framework for generating and analyzing smFISH data in complex tissues while overcoming autofluorescence and increasing multiplexing capacity. We developed dotdotdot (https://github.com/LieberInstitute/dotdotdot) as a corresponding software package to quantify RNA transcripts in single nuclei and perform differential expression analysis. We first demonstrate robustness of our platform in single mouse neurons by quantifying differential expression of activity-regulated genes. We then quantify spatial gene expression in human dorsolateral prefrontal cortex (DLPFC) using spectral imaging and dotdotdot to mask lipofuscin autofluorescence. We lastly apply machine learning to predict cell types and perform downstream cell type-specific expression analysis. In summary, we provide experimental workflows, imaging acquisition and analytic strategies for quantification and biological interpretation of smFISH data in complex tissues.}",
+    issn = {0305-1048},
+    doi = {10.1093/nar/gkaa312},
+    url = {https://doi.org/10.1093/nar/gkaa312},
+    eprint = {https://academic.oup.com/nar/article-pdf/48/11/e66/33384812/gkaa312.pdf},
+}
