### 2020-11-13 3f9c7bc
adds references to other tools. fixes #1905
