### 2020-11-13 21c3028
clarifies required skills and knowledge in paper
