+++ paper/paper.md
@@ -73,6 +73,7 @@ Furthermore, *starfish* implements comprehensive logging of all data processing 
 
 There are a number of other tools which support localization and quantification of spots in fluorescent microscopy images, including ImageJ and CellProfiler, however these tools do not support multiplexed decoding of gene targets necessary for many assays. Other tools which are designed for more specifically to handle the kinds of assays that starfish supports include [dotdotdot](https://github.com/LieberInstitute/dotdotdot) (@dotdotdot),  [pysmFISH](https://github.com/linnarsson-lab/pysmFISH/), and [SMART-Q](https://github.com/shenlab-ucsf/SMART-Q) (forked from an earlier development release of starfish) (@SMART-Q).
 
+starfish requires a working knowledge of Python and fluorescent image analysis for a user to create an analysis pipeline.
 To help new users get started and support the broader single cell biology community in learning how to work with these data, *starfish* maintains example datasets and reference implementations ported from published assays, including
 MERFISH (@MERFISH),
 In Situ Sequencing (@ISS),
@@ -82,6 +83,7 @@ smFISH (@smFISH),
 DARTFISH (@DARTFISH),
 STARmap (@starmap),
 and seqFISH (@seqFISH).
+To take advantage of starfish's support for large scale processing, users must have familiarity with cluster or cloud computing.
 
 *starfish* was developed alongside the [SpaceTx project](https://spacetx-starfish.readthedocs.io/en/stable/about/index.html), a CZI-funded effort to compare spatial transcriptomics methods in the context of determining cell types in the brain (@starfish-sfn).
 *starfish* is currently in use by multiple research groups, including the [Allen Institute for Brain Science](https://alleninstitute.org), the [Chan Zuckerberg Biohub](https://www.czbiohub.org/), and the [Zhang Lab at UC San Diego](http://jinzhanglab.ucsd.edu/).
