### 2020-11-13 f2e3a1a
clarifies skills in docs
