### 2020-11-13 0623a6a
adds clarifications about the purpose of formatting examples
one more
