+++ examples/data_formatting_examples/README.txt
@@ -3,5 +3,11 @@
 Data Formatting Examples Using TileFetcher
 ------------------------------------------
 
-Several examples of how to format data into SpaceTx format from a variety of
-types of input data shapes and sizes, demonstrating the flexibility of :py:class:`.TileFetcher`.
+There is currently no widely used standard data format for image-based microscopy data.
+starfish works best when data is converted into the SpaceTx data format.
+Here, we provide several examples of how to convert data from non-standardized data formats
+into SpaceTx format from a variety of types of input data shapes and sizes, demonstrating the
+flexibility of :py:class:`.TileFetcher`.
+
+These examples are provided for illustrative purposes, demonstrating how the
+:py:class:`.TileFetcher` is used in practice. They will need to be adapted to meet the specific needs of your data.
