### 2020-11-13 d2ade82
adds clarifications about the purpose of formatting examples
one more
