### 2020-11-13 e08044a
adds seaborn as a prereq for the quickstart
