### 2020-11-13 035d6cd
adds additional sections to user guide
