### 2020-11-13 spacetx/starfish#1907 @neuromusic
> Thanks for writing this software, I think the starfish & its associated documentation represents a huge amount of work and the authors as well as all involved should be commended for their efforts!

Thank you! It's been a big effort from multiple folks!!

>Since other data modalities are also measured (e.g. protein), it's worth using more general wording than "gene expression", perhaps "abundance" is more relevant.

starfish is designed to localize and decode gene expression from mRNA spots and does not currently support quantification of protein abundance.

> It's difficult to assess from the JOSS MS and the software documentation who exactly is the target audience.

This is a good point. I've added clarifications in the manuscript and user guide to clarify expectations of our user's knowledge.

> How do you ensure that various file format updates for technologies are dealt with? Is it within scope of starfish developers to provide scripts for different versions of input files? I think answers to these should be mentioned in the Tutorials and Examples area.

Most of the file formats produced by the various SpaceTx labs are not standard, but are ad hoc implementations. We do not commit to supporting these file formats and the formatting examples are only provided as examples of how to convert data to the spacetx format, using starfish's TileFetcher to perform the conversion. I've added notes to the Tutorials and Examples to help clarify.

> mentioning [seaborn is] needed at the beginning of the Quick start tutorial

nice catch. done.

> This is actually given in "Working with Starfish outputs" under "Help & Reference", but IMO could appear in the Tutorials and Examples area for greater visibility.

> it's worth just making the "Help & Reference" section more visible, esp the 'Processing with AWS' section.

Hmm... yeah, part of this is purely a pragmatic limitation of sphinx: Everything in Examples and Tutorials is a python script which can be downloaded and run by the use, while the "output" examples and AWS tutorial are non-executable. Nonetheless, I've moved these resources into the user guide and added sections highlighting these resources.

Thank you @shazanfar for your detailed review!!
