### 2020-11-13 #2440 @neuromusic
Thank you for your patience, @csoneson and all.

I've addressed the outstanding issues from @vals and @shazanfar in [PR #1913](https://github.com/spacetx/starfish/pull/1913), except for the request for rendered examples in [starfish#1904](https://github.com/spacetx/starfish/issues/1904), which I've addressed in a comment on that issue.

As soon as the tests pass and my colleague @mattcai reviews the PR, these changes will be live in the manuscript and documentation.
