### 2020-11-13 dbd0a53
fix "blank line contains white space" error on sphinx build
