### 2020-11-13 154364c
remove dead references
