+++ docs/source/help_and_reference/index.rst
@@ -9,8 +9,6 @@ Help & Reference
     data_model/index
     spacetx-format/index
     available_datasets/index
-    working_with_starfish_outputs/index
-    processing_at_scale/index
     configuration/index
     glossary/glossary
     spacetx_consortium/index
