+++ docs/source/user_guide/processing_at_scale/index.rst
@@ -1,4 +1,5 @@
-===================
+.. _processing_at_scale:
+
 Processing With AWS
 ===================
 
