### 2020-11-14 7001910
Fix link to seaborn in quick start tutorial
