### 2020-11-15 #2440 @csoneson
Thanks @neuromusic. I have removed the `paused` label. Once the changes are merged you can ask whedon to re-render the manuscript: `@whedon generate pdf`, and @giovp, @shazanfar, @vals can have a look at the revisions.
