+++ starfish/REQUIREMENTS-STRICT.txt
@@ -38,7 +38,7 @@ mpmath==1.1.0
 nbconvert==5.6.1
 nbformat==5.0.4
 networkx==2.4
-notebook==6.0.3
+notebook==6.1.5
 numpy==1.18.2
 packaging==20.3
 pandas==1.0.3
