### 2020-11-18 732b9e0
Refresh requirements.
Periodic refresh of requirements prompted by https://github.com/spacetx/starfish/pull/1914/

Test plan: -alltest should cover all tests.
