+++ requirements/REQUIREMENTS-CI.txt
@@ -2,110 +2,120 @@
 alabaster==0.7.12
 apipkg==1.5
 appnope==0.1.0
-attrs==19.3.0
-Babel==2.8.0
-backcall==0.1.0
-bleach==3.1.4
-boto3==1.12.32
-botocore==1.15.32
-certifi==2019.11.28
+argon2-cffi==20.1.0
+async-generator==1.10
+attrs==20.3.0
+Babel==2.9.0
+backcall==0.2.0
+bleach==3.2.1
+boto3==1.16.21
+botocore==1.19.21
+cached-property==1.5.2
+certifi==2020.11.8
+cffi==1.14.3
 chardet==3.0.4
-click==7.1.1
+click==7.1.2
+colorama==0.4.4
 commonmark==0.9.1
-coverage==5.0.4
+coverage==5.3
 cycler==0.10.0
 dataclasses==0.6
 decorator==4.4.2
 defusedxml==0.6.0
-diskcache==4.1.0
-docutils==0.15.2
+diskcache==5.1.0
+docutils==0.16
 entrypoints==0.3
 execnet==1.7.1
-flake8==3.7.9
+flake8==3.8.4
 flake8-import-order==0.18.1
-h5py==2.10.0
-idna==2.9
-imageio==2.8.0
+h5py==3.1.0
+idna==2.10
+imageio==2.9.0
 imagesize==1.2.0
-importlib-metadata==1.6.0
-ipykernel==5.2.0
-ipython==7.13.0
+importlib-metadata==2.0.0
+iniconfig==1.1.1
+ipykernel==5.3.4
+ipython==7.16.1
 ipython-genutils==0.2.0
 ipywidgets==7.5.1
-jedi==0.16.0
-Jinja2==2.11.1
-jmespath==0.9.5
-joblib==0.14.1
+jedi==0.17.2
+Jinja2==2.11.2
+jmespath==0.10.0
+joblib==0.17.0
 jsonschema==3.2.0
-jupyter-client==6.1.2
-jupyter-core==4.6.3
-keyring==21.2.0
-kiwisolver==1.1.0
+jupyter-client==6.1.7
+jupyter-core==4.7.0
+jupyterlab-pygments==0.1.2
+keyring==21.5.0
+kiwisolver==1.3.1
 m2r==0.2.1
 MarkupSafe==1.1.1
-matplotlib==3.2.1
+matplotlib==3.3.3
 mccabe==0.6.1
 mistune==0.8.4
-more-itertools==8.2.0
 mpmath==1.1.0
-mypy==0.770
+mypy==0.790
 mypy-extensions==0.4.3
-nbconvert==5.6.1
+nbclient==0.5.1
+nbconvert==6.0.7
 nbencdec==0.0.7
-nbformat==5.0.4
-networkx==2.4
-notebook==6.0.3
-numpy==1.18.2
-numpydoc==0.9.2
-packaging==20.3
-pandas==1.0.3
-pandocfilters==1.4.2
-parso==0.6.2
+nbformat==5.0.8
+nest-asyncio==1.4.3
+networkx==2.5
+notebook==6.1.5
+numpy==1.19.4
+numpydoc==1.1.0
+packaging==20.4
+pandas==1.1.4
+pandocfilters==1.4.3
+parso==0.7.1
 pexpect==4.8.0
 pickleshare==0.7.5
-Pillow==7.0.0
-pkginfo==1.5.0.1
+Pillow==8.0.1
+pkginfo==1.6.1
 pluggy==0.13.1
-prometheus-client==0.7.1
-prompt-toolkit==3.0.5
+prometheus-client==0.9.0
+prompt-toolkit==3.0.8
[…]

[…]
