+++ requirements/REQUIREMENTS-NAPARI-CI.txt
@@ -50,7 +50,7 @@ napari-svg==0.1.3
 nbconvert==5.6.1
 nbformat==5.0.4
 networkx==2.4
-notebook==6.0.3
+notebook==6.1.5
 numpy==1.18.2
 numpydoc==0.9.2
 packaging==20.3
