### 2020-11-18 017f796
Bump pillow from 7.0.0 to 7.1.0 in /starfish
Bumps [pillow](https://github.com/python-pillow/Pillow) from 7.0.0 to 7.1.0.
- [Release notes](https://github.com/python-pillow/Pillow/releases)
- [Changelog](https://github.com/python-pillow/Pillow/blob/master/CHANGES.rst)
- [Commits](https://github.com/python-pillow/Pillow/compare/7.0.0...7.1.0)

Signed-off-by: dependabot[bot] <support@github.com>
