### 2020-11-19 703a0d9
Use python requests to bypass certificate errors
On travis's windows machines, the certs are not set up correctly.  pandas uses python's default url library, which uses the system setup.  python requests, however, uses the mozilla root cert store, which is independent of the system config.
