### 2020-11-21 84f16a0
Move to pillow 7.1.2 instead.
