+++ starfish/REQUIREMENTS-STRICT.txt
@@ -46,7 +46,7 @@ pandocfilters==1.4.2
 parso==0.6.2
 pexpect==4.8.0
 pickleshare==0.7.5
-Pillow==7.1.0
+Pillow==7.1.2
 prometheus-client==0.7.1
 prompt-toolkit==3.0.5
 ptyprocess==0.6.0
