### 2020-11-21 4860e41
Move to pillow 7.1.2 instead.
See https://github.com/spacetx/starfish/pull/1918#issuecomment-731504327
