+++ requirements/REQUIREMENTS-NAPARI-CI.txt
@@ -59,7 +59,7 @@ pandocfilters==1.4.2
 parso==0.6.2
 pexpect==4.8.0
 pickleshare==0.7.5
-Pillow==7.1.0
+Pillow==7.1.2
 pluggy==0.13.1
 prometheus-client==0.7.1
 prompt-toolkit==3.0.5
