+++ requirements/REQUIREMENTS-CI.txt
@@ -63,7 +63,7 @@ pandocfilters==1.4.2
 parso==0.6.2
 pexpect==4.8.0
 pickleshare==0.7.5
-Pillow==7.1.0
+Pillow==7.1.2
 pkginfo==1.5.0.1
 pluggy==0.13.1
 prometheus-client==0.7.1
