### 2020-11-30 2412863
Ignore requirements that are already satisfied
Don't force-reinstall as that can cause conflicts with environments with distutils-installed packages.

Fixes #1922
