### 2020-12-04 spacetx/starfish#1925 @vals
Tutorials making use of Napari should include screenshots of expected output.
[ Related to openjournals/joss-reviews#2440 ]

The tutorials should include screenshots of the expected Napari window when this feature is demonstrated. E.g. https://spacetx-starfish.readthedocs.io/en/latest/gallery/quick_start/plot_quick_start.html#sphx-glr-gallery-quick-start-plot-quick-start-py or https://spacetx-starfish.readthedocs.io/en/latest/gallery/pipelines/3d_smFISH.html#load-data-run-pipeline-display-results or the bottom section of https://spacetx-starfish.readthedocs.io/en/latest/gallery/how_to/plot_display.html

For https://spacetx-starfish.readthedocs.io/en/latest/gallery/how_to/plot_display.html in particular, it would be good to annotate the screenshots with e.g. arrows pointing to the ‘stack [1]’ layer. Similar should also be done in other potential tutorials demonstrating building on the interactive visualization.
