### 2020-12-04 spacetx/starfish#1926 @vals
Are the masks in the left in the tutorial really two cells?
[ Related to openjournals/joss-reviews/issues/2440 ]

Are the masks in the left of the example in the tutorial two cells? There seems to be two nuclei, but are they two cells? Maybe a different less confusing FOV could be used? Or this could be commented upon in the segmentation section of the Quick Start tutorial.

![image](https://user-images.githubusercontent.com/668803/101120385-80bda200-35bb-11eb-9150-d9be8e8579b1.png)
