### 2020-12-04 spacetx/starfish#1927 @vals
JOSS paper should mention how starfish compares to the other available tools
[ Related to https://github.com/openjournals/joss-reviews/issues/2440 ]

In the new revision the JOSS manuscript mentions dotdotdot, pysmFISH, and SMART-Q. It should also be pointed out how starfish differ from these tools so a potential user can make an informed choice.
