### 2020-12-04 #2440 @vals
Hi again @csoneson , I have gone over the checklist again now for a second round.

There are still some issues regarding the documentation (linked above). Additionally, the manuscript doesn't mention how starfish differ from the alternative tools mentioned. I'm sure the authors could write some comments on how flexible and scalable starfish is compared to alternative tools.

Best,
/Valentine
