### 2020-12-04 #2440 @csoneson
Thanks @vals - @neuromusic, could you take a look at expanding the documentation a bit and highlighting how starfish compares to other tools, as indicated above?
