### 2020-12-09 #2440 @neuromusic
I think we can address most of them, however, I need to look into https://github.com/spacetx/starfish/issues/1925 more carefully... I've explored automatically capturing napari screenshots, but I'm unsure of how easy it will be to do as part of our documentation build process.
