### 2020-12-09 spacetx/starfish#1924 @neuromusic
(edited my comment... the last line should give the napari version)
