### 2020-12-09 #2440 @vals
Do they need to be captured automatically? There are other screenshots in the tutorials. The examples I gave are actually exceptions. It is hard to follow what is going on with the Napari window if it's not pointed out with images in the documentation.
