### 2020-12-10 137a012
Fixed Bug. Added features dimension to intensity table when created using TraceBuildingStrategies.NEAREST_NEIGHBOR.
