### 2020-12-10 0998ed6
Needed to add spot_id to intensity table in order to AssignSpots.
