### 2020-12-10 #2440 @neuromusic
@vals yes, its important to us that all documentation involving data can be automatically generated so that the images don't become stale if the tutorial gets updated in the future. all of the images in the other tutorials are generated through matplotlib, etc.

let me look into this a bit more. I've got a PR open where I started trying to capture napari screenshots programmatically, which I need to revisit. It also might make sense to simply use matplotlib for the quick start instead of napari
