### 2020-12-13 spacetx/starfish#1924 @vals
Hi @neuromusic, I just reran the same notebook as I used when reviewing. This time it didn't crash! (It is very likely I updated something in my Python environment between then and now).

I am pasting the outputs of the version numbers for the sake of completeness:

```
Windows-10-10.0.19041-SP0
Python 3.7.4 (default, Aug  9 2019, 18:34:13) [MSC v.1915 64 bit (AMD64)]
NumPy 1.18.1
SciPy 1.4.1
scikit-image 0.15.0
pandas 1.0.3
sklearn 0.22.2.post1
xarray 0.16.0
sympy 1.4
starfish 0.2.1
napari 0.3.7
```

Since it all seems to be working now, I am closing the issue.
