### 2021-01-03 fe512dd
Add necessary coords for IntensityTable when using Nearest Neighbors strategy (#1928)
* Fixed Bug. Added features dimension to intensity table when created using TraceBuildingStrategies.NEAREST_NEIGHBOR.

* Needed to add spot_id to intensity table in order to AssignSpots.
