### 2021-01-14 0ec2bdd
resolve match_histograms in higher version of skimage
