### 2021-01-14 2ce4cdd
change version resolving order
