### 2021-01-25 #2440 @neuromusic
Thanks, I'll provide an update on the outstanding item shortly.

On Fri, Jan 22, 2021, 07:04 Charlotte Soneson <notifications@github.com>
wrote:

> @neuromusic <https://github.com/neuromusic> - please let us know how this
> is progressing. It looks like we are pretty close to the finishing line,
> but at this point I would like to ask for a timeline of when you expect to
> be able to address the remaining comments, or whether there are issues that
> need to be discussed.
>
