### 2021-01-31 #2440 @neuromusic
quick update: I've got screenshot rendering working locally. I just need to get it working on CI & readthedocs. Hoping to get that taken care of this week.
