+++ starfish/REQUIREMENTS-STRICT.txt
@@ -2,7 +2,7 @@
 appnope==0.1.0
 attrs==19.3.0
 backcall==0.1.0
-bleach==3.1.4
+bleach==3.3.0
 boto3==1.12.32
 botocore==1.15.32
 certifi==2019.11.28
