+++ requirements/REQUIREMENTS-NAPARI-CI.txt
@@ -4,7 +4,7 @@ appnope==0.1.0
 attrs==19.3.0
 Babel==2.8.0
 backcall==0.1.0
-bleach==3.1.4
+bleach==3.3.0
 boto3==1.12.32
 botocore==1.15.32
 cachey==0.2.1
