### 2021-02-17 4aa5e22
Import match_histograms from skimage.exposure
It was [moved to skimage.exposure @ 0.16.0](https://scikit-image.org/docs/0.16.x/release_notes_and_installation.html#deprecations)

Fixes: #1932

Test plan: Travis
