### 2021-02-17 9179692
additional type hacks
