+++ .travis.yml
@@ -64,6 +64,7 @@ jobs:
         - sudo apt-get install -y libgl1-mesa-glx libqt5x11extras5 xvfb
         - make install-dev
         - pip install .[napari] pytest-qt
+        - pip freeze
         - export DISPLAY=:99
         - Xvfb $DISPLAY -ac -screen 0 1024x768x24 &
         - sleep 10
@@ -78,6 +79,7 @@ jobs:
         - pip install -U pip==21.0.1
         - pip install -r requirements/REQUIREMENTS-NAPARI-CI.txt
         - make install-dev
+        - pip freeze
         - export DISPLAY=:99
         - Xvfb $DISPLAY -ac -screen 0 1024x768x24 &
         - sleep 10
