+++ requirements/REQUIREMENTS-NAPARI-CI.txt
@@ -1,100 +1,110 @@
 # You should not edit this file directly.  Instead, you should edit one of the following files (requirements/REQUIREMENTS-NAPARI-CI.txt.in) and run make requirements/REQUIREMENTS-NAPARI-CI.txt
 alabaster==0.7.12
-appnope==0.1.0
-attrs==19.3.0
-Babel==2.8.0
-backcall==0.1.0
+appdirs==1.4.4
+appnope==0.1.2
+argon2-cffi==20.1.0
+async-generator==1.10
+attrs==20.3.0
+Babel==2.9.0
+backcall==0.2.0
 bleach==3.3.0
-boto3==1.12.32
-botocore==1.15.32
+boto3==1.17.10
+botocore==1.20.10
+cached-property==1.5.2
 cachey==0.2.1
-certifi==2019.11.28
-chardet==3.0.4
-click==7.1.1
+certifi==2020.12.5
+cffi==1.14.5
+chardet==4.0.0
+click==7.1.2
 cycler==0.10.0
-dask==2.13.0
+dask==2021.2.0
 dataclasses==0.6
 decorator==4.4.2
 defusedxml==0.6.0
-diskcache==4.1.0
-docutils==0.15.2
+diskcache==5.2.1
+docutils==0.16
 entrypoints==0.3
-freetype-py==2.1.0.post1
-fsspec==0.7.1
-h5py==2.10.0
+freetype-py==2.2.0
+h5py==3.1.0
 HeapDict==1.0.1
-idna==2.9
-imageio==2.8.0
+idna==2.10
+imageio==2.9.0
 imagesize==1.2.0
-importlib-metadata==1.6.0
-ipykernel==5.2.0
-ipython==7.13.0
+importlib-metadata==3.4.0
+iniconfig==1.1.1
+ipykernel==5.4.3
+ipython==7.16.1
 ipython-genutils==0.2.0
-ipywidgets==7.5.1
-jedi==0.16.0
-Jinja2==2.11.1
-jmespath==0.9.5
-joblib==0.14.1
+ipywidgets==7.6.3
+jedi==0.18.0
+Jinja2==2.11.3
+jmespath==0.10.0
+joblib==1.0.1
 jsonschema==3.2.0
-jupyter-client==5.3.4
-jupyter-core==4.6.3
-kiwisolver==1.1.0
+jupyter-client==6.1.11
+jupyter-core==4.7.1
+jupyterlab-pygments==0.1.2
+jupyterlab-widgets==1.0.0
+kiwisolver==1.3.1
 MarkupSafe==1.1.1
-matplotlib==3.2.1
+matplotlib==3.3.4
 mistune==0.8.4
-more-itertools==8.2.0
-mpmath==1.1.0
-napari==0.3.4
-napari-plugin-engine==0.1.6
-napari-svg==0.1.3
-nbconvert==5.6.1
-nbformat==5.0.4
-networkx==2.4
-notebook==6.1.5
-numpy==1.18.2
-numpydoc==0.9.2
-packaging==20.3
-pandas==1.0.3
-pandocfilters==1.4.2
-parso==0.6.2
+mpmath==1.2.1
+napari==0.3.8
+napari-plugin-engine==0.1.9
+napari-svg==0.1.4
+nbclient==0.5.2
+nbconvert==6.0.7
+nbformat==5.1.2
+nest-asyncio==1.5.1
+networkx==2.5
+notebook==6.2.0
+numpy==1.19.5
+numpydoc==1.1.0
+packaging==20.9
+pandas==1.1.5
+pandocfilters==1.4.3
[…]
