+++ requirements/REQUIREMENTS-CI.txt.in
@@ -1,7 +1,7 @@
 # requirements to run CI except for napari
 flake8
 flake8-import-order
-m2r
+m2r2
 mypy != 0.740
 numpydoc
 nbencdec >= 0.0.5
