### 2021-02-18 3f45bc4
Update dependencies.
- Some style changes are required for the new linter versions.
- New version of napari requires explicit installation of PySide2 or PyQT5
