+++ starfish/core/image/Filter/map.py
@@ -94,8 +94,8 @@ class Map(FilterAlgorithm):
         elif isinstance(func, FunctionSourceBundle):
             if module is not None:
                 raise ValueError(
-                    f"When passing in the function as a `FunctionSourceBundle`, module should not "
-                    f"be set."
+                    "When passing in the function as a `FunctionSourceBundle`, module should not "
+                    "be set."
                 )
             self.func = func
         self.in_place = in_place
