+++ requirements/REQUIREMENTS-CI.txt
@@ -1,114 +1,123 @@
 # You should not edit this file directly.  Instead, you should edit one of the following files (requirements/REQUIREMENTS-CI.txt.in) and run make requirements/REQUIREMENTS-CI.txt
 alabaster==0.7.12
 apipkg==1.5
-appnope==0.1.0
+appnope==0.1.2
 argon2-cffi==20.1.0
-attrs==19.3.0
-Babel==2.8.0
-backcall==0.1.0
+async-generator==1.10
+attrs==20.3.0
+Babel==2.9.0
+backcall==0.2.0
 bleach==3.3.0
-boto3==1.12.32
-botocore==1.15.32
-certifi==2019.11.28
+boto3==1.17.10
+botocore==1.20.10
+cached-property==1.5.2
+certifi==2020.12.5
 cffi==1.14.5
-chardet==3.0.4
-click==7.1.1
+chardet==4.0.0
+click==7.1.2
+colorama==0.4.4
 commonmark==0.9.1
-coverage==5.0.4
+coverage==5.4
 cycler==0.10.0
 dataclasses==0.6
 decorator==4.4.2
 defusedxml==0.6.0
-diskcache==4.1.0
-docutils==0.15.2
+diskcache==5.2.1
+docutils==0.16
 entrypoints==0.3
-execnet==1.7.1
-flake8==3.7.9
+execnet==1.8.0
+flake8==3.8.4
 flake8-import-order==0.18.1
-h5py==2.10.0
-idna==2.9
-imageio==2.8.0
+h5py==3.1.0
+idna==2.10
+imageio==2.9.0
 imagesize==1.2.0
-importlib-metadata==1.6.0
-ipykernel==5.2.0
-ipython==7.13.0
+importlib-metadata==3.4.0
+iniconfig==1.1.1
+ipykernel==5.4.3
+ipython==7.16.1
 ipython-genutils==0.2.0
-ipywidgets==7.5.1
-jedi==0.16.0
-Jinja2==2.11.1
-jmespath==0.9.5
-joblib==0.14.1
+ipywidgets==7.6.3
+jedi==0.18.0
+Jinja2==2.11.3
+jmespath==0.10.0
+joblib==1.0.1
 jsonschema==3.2.0
-jupyter-client==6.1.2
-jupyter-core==4.6.3
-keyring==21.2.0
-kiwisolver==1.1.0
+jupyter-client==6.1.11
+jupyter-core==4.7.1
+jupyterlab-pygments==0.1.2
+jupyterlab-widgets==1.0.0
+keyring==22.0.1
+kiwisolver==1.3.1
 m2r2==0.2.7
 MarkupSafe==1.1.1
-matplotlib==3.2.1
+matplotlib==3.3.4
 mccabe==0.6.1
 mistune==0.8.4
-more-itertools==8.2.0
-mpmath==1.1.0
-mypy==0.770
+mpmath==1.2.1
+mypy==0.800
 mypy-extensions==0.4.3
-nbconvert==5.6.1
-nbencdec==0.0.7
-nbformat==5.0.4
-networkx==2.4
-notebook==6.1.5
-numpy==1.18.2
-numpydoc==0.9.2
-packaging==20.3
-pandas==1.0.3
-pandocfilters==1.4.2
-parso==0.6.2
+nbclient==0.5.2
+nbconvert==6.0.7
+nbencdec==0.0.10
+nbformat==5.1.2
+nest-asyncio==1.5.1
+networkx==2.5
+notebook==6.2.0
+numpy==1.19.5
+numpydoc==1.1.0
+packaging==20.9
+pandas==1.1.5
[…]

[…]
