### 2021-02-18 0af3521
Update dependencies.
- Some style changes are required for the new linter versions.
- New version of napari requires explicit installation of PySide2 or PyQT5
