+++ starfish/core/morphology/Filter/map.py
@@ -62,8 +62,8 @@ class Map(FilterAlgorithm):
         elif isinstance(func, FunctionSourceBundle):
             if module is not None:
                 raise ValueError(
-                    f"When passing in the function as a `FunctionSourceBundle`, module should not "
-                    f"be set."
+                    "When passing in the function as a `FunctionSourceBundle`, module should not "
+                    "be set."
                 )
             self._func = func
         self._func_args = func_args
