### 2021-02-19 9902753
add napari to quickstart
