### 2021-03-14 spacetx/starfish#1926 @neuromusic
Indeed, they are!

As noted in the quick-start, the cells in this dataset express either the mouse or human ACTB gene, but not both. This can be seen by exploring each of the images in the stack, where you can see that, despite the apparent ambiguity, the segmentation has distinguished between two cells... one expressing the human ACTB and one expressing the mouse ACTB.
