### 2021-03-14 5f129d7
clarified segmentation results. fixes #1926
