### 2021-03-14 4ebe24a
clarified segmentation results. fixes #1926
