### 2021-03-14 spacetx/starfish#1926 @neuromusic
I've added a clarification to the quickstart.
