### 2021-03-14 #2440 @neuromusic
Thank you @csoneson! I have a PR open to address the outstanding issues. 2 of the 3 are fixed and I will have a solution for the final issue within the week (by Friday, 3/19).

https://github.com/spacetx/starfish/pull/1938
