### 2021-03-14 87ff56d
remove commented-out display code
