### 2021-03-14 spacetx/starfish#1925 @neuromusic
As noted in the discussion on openjournals/joss-reviews#2440, we've strived to make our documentation fully-automated. However, incorporating napari into the documentation building as resulted in [failures on ReadTheDocs that we weren't able to debug](https://github.com/readthedocs/readthedocs.org/issues/7942).

I have removed references to the interactive display utility from the example pipelines in #1938 and will replace the interactive visualizations in the Quick Start with matplotlib-based visualizations.
