+++ starfish/REQUIREMENTS-STRICT.txt
@@ -53,7 +53,7 @@ pandocfilters==1.4.3
 parso==0.8.1
 pexpect==4.8.0
 pickleshare==0.7.5
-Pillow==8.1.0
+Pillow==8.1.1
 prometheus-client==0.9.0
 prompt-toolkit==3.0.16
 ptyprocess==0.7.0
