+++ requirements/REQUIREMENTS-CI.txt
@@ -72,7 +72,7 @@ pandocfilters==1.4.3
 parso==0.8.1
 pexpect==4.8.0
 pickleshare==0.7.5
-Pillow==8.1.0
+Pillow==8.1.1
 pkginfo==1.7.0
 pluggy==0.13.1
 prometheus-client==0.9.0
