+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @Article{Seurat,
   author = {Tim Stuart and Andrew Butler and Paul Hoffman and
-    Christoph Hafemeister and Efthymia Papalexi and William M Mauck
-    III and Yuhan Hao and Marlon Stoeckius and Peter Smibert and
+    Christoph Hafemeister and Efthymia Papalexi and William M {Mauck
+    III} and Yuhan Hao and Marlon Stoeckius and Peter Smibert and
     Rahul Satija},
   title = {Comprehensive Integration of Single-Cell Data},
   journal = {Cell},
@@ -36,7 +36,7 @@
   number = {2},
   pages = {115--121},
   author = {Wolfgang Huber and Vincent J Carey and Robert Gentleman and Simon Anders and Marc Carlson and Benilton S Carvalho and Hector Corrada Bravo and Sean Davis and Laurent Gatto and Thomas Girke and Raphael Gottardo and Florian Hahne and Kasper D Hansen and Rafael A Irizarry and Michael Lawrence and Michael I Love and James MacDonald and Valerie Obenchain and Andrzej K Ole{\'{s}} and Herv{\'{e}} Pag{\`{e}}s and Alejandro Reyes and Paul Shannon and Gordon K Smyth and Dan Tenenbaum and Levi Waldron and Martin Morgan},
-  title = {Orchestrating high-throughput genomic analysis with Bioconductor},
+  title = {Orchestrating high-throughput genomic analysis with {B}ioconductor},
   journal = {Nature Methods}
 }
 
@@ -90,7 +90,7 @@
   volume = {15},
   number = {11},
   pages = {932--935},
-  author = {Simone Codeluppi and Lars E. Borm and Amit Zeisel and Gioele La Manno and Josina A. van Lunteren and Camilla I. Svensson and Sten Linnarsson},
+  author = {Simone Codeluppi and Lars E. Borm and Amit Zeisel and Gioele La Manno and Josina A. {van Lunteren} and Camilla I. Svensson and Sten Linnarsson},
   title = {Spatial organization of the somatosensory cortex revealed by {osmFISH}},
   journal = {Nature Methods}
 }
@@ -137,7 +137,7 @@
       author = {van der Walt, {S}t\'efan and {S}ch\"onberger, {J}ohannes {L}. and
                 {Nunez-Iglesias}, {J}uan and {B}oulogne, {F}ran\c{c}ois and {W}arner,
                 {J}oshua {D}. and {Y}ager, {N}eil and {G}ouillart, {E}mmanuelle and
-                {Y}u, {T}ony and the scikit-image contributors},
+                {Y}u, {T}ony and {the scikit-image contributors}},
       year = {2014},
       month = {6},
       keywords = {Image processing, Reproducible research, Education,
@@ -226,7 +226,7 @@
 }
 
 @article{SMART-Q,
-    author = {Yang, Xiaoyu AND Bergenholtz, Seth AND Maliskova, Lenka AND Pebworth, Mark-Phillip AND Kriegstein, Arnold R. AND Li, Yun AND Shen, Yin},
+    author = {Yang, Xiaoyu and Bergenholtz, Seth and Maliskova, Lenka and Pebworth, Mark-Phillip and Kriegstein, Arnold R. and Li, Yun and Shen, Yin},
     journal = {PLOS ONE},
     publisher = {Public Library of Science},
     title = {SMART-Q: An Integrative Pipeline Quantifying Cell Type-Specific RNA Transcription},
@@ -246,7 +246,7 @@
     journal = {Nucleic Acids Research},
     volume = {48},
     number = {11},
-    pages = {e66-e66},
+    pages = {e66},
     year = {2020},
     month = {05},
[…]
