+++ paper/paper.md
@@ -65,27 +65,27 @@ Furthermore, many of these assays require specialized analysis to decode gene ex
 It lets biologists build scalable pipelines that localize and quantify RNA transcripts in image data generated by any hybridization- or sequencing-based *in situ* transcriptomics method, from classic RNA single-molecule FISH to combinatorial barcoded assays.
 Image processing of an experiment is divided into fields of view (FOV) that correspond to the data produced by a microscope at a single location on a microscope slide.
 *starfish* lets users register and pre-process images in each FOV, localize spots representing tagged RNA molecules in 3D, decode the identity of those molecules according to the experimental design, segment cells, assign the spots to cells, then aggregate spots into a cell x gene expression matrix.
-This spatially-annotated gene expression matrix can then be analyzed and visualized in downstream tools for single-cell biology, such as Seurat (@Seurat), Bioconductor (@Bioconductor), ScanPy (@ScanPy), and cellxgene (@cellxgene).
+This spatially-annotated gene expression matrix can then be analyzed and visualized in downstream tools for single-cell biology, such as Seurat [@Seurat], Bioconductor [@Bioconductor], Scanpy [@ScanPy], and cellxgene [@cellxgene].
 
 To enable large scale processing of these data, *starfish* leverages a 5-dimensional imaging data model (x, y, z, round, channel) backed by the cloud-friendly `spacetx-format` file format, and [slicedimage](https://github.com/spacetx/slicedimage/), an interface for lazy, distributed loading of `spacetx-format` datasets.  
 Furthermore, *starfish* implements comprehensive logging of all data processing steps for provenance tracking, reproducibility, and transparency.
-*starfish* is built on top of popular Python tools like xarray (@xarray) and scikit-image (@scikit-image).
+*starfish* is built on top of popular Python tools like xarray [@xarray] and scikit-image [@scikit-image].
 
-There are a number of other tools which support localization and quantification of spots in fluorescent microscopy images, including ImageJ and CellProfiler, however these tools do not support multiplexed decoding of gene targets necessary for many assays. Other tools which are designed for more specifically to handle the kinds of assays that starfish supports include [dotdotdot](https://github.com/LieberInstitute/dotdotdot) (@dotdotdot), a MATLAB toolbox designed for RNAscope assays;  [pysmFISH](https://github.com/linnarsson-lab/pysmFISH/), a Python package designed for smFISH assays; and [SMART-Q](https://github.com/shenlab-ucsf/SMART-Q) fork from an earlier development release of starfish and adding support for immunostaining and other features (@SMART-Q).
+There are a number of other tools which support localization and quantification of spots in fluorescent microscopy images, including ImageJ and CellProfiler, however these tools do not support multiplexed decoding of gene targets necessary for many assays. Other tools which are designed for more specifically to handle the kinds of assays that starfish supports include [dotdotdot](https://github.com/LieberInstitute/dotdotdot) [@dotdotdot], a MATLAB toolbox designed for RNAscope assays;  [pysmFISH](https://github.com/linnarsson-lab/pysmFISH/), a Python package designed for smFISH assays; and [SMART-Q](https://github.com/shenlab-ucsf/SMART-Q) fork from an earlier development release of starfish and adding support for immunostaining and other features [@SMART-Q].
 
 starfish requires a working knowledge of Python and fluorescent image analysis for a user to create an analysis pipeline.
[…]
