+++ paper/paper.bib
@@ -103,7 +103,7 @@
   publisher = {Oxford University Press ({OUP})},
   volume = {46},
   number = {4},
-  pages = {e22--e22},
+  pages = {e22},
   author = {Xiaoyin Chen and Yu-Chi Sun and George M Church and Je Hyuk Lee and Anthony M Zador},
   title = {Efficient in situ barcode sequencing using padlock probe-based {BaristaSeq}},
   journal = {Nucleic Acids Research}
@@ -134,7 +134,7 @@
 
 @article{scikit-image,
       title = {scikit-image: image processing in {P}ython},
-      author = {van der Walt, {S}t\'efan and {S}ch\"onberger, {J}ohannes {L}. and
+      author = {{van der Walt}, {S}t\'efan and {S}ch\"onberger, {J}ohannes {L}. and
                 {Nunez-Iglesias}, {J}uan and {B}oulogne, {F}ran\c{c}ois and {W}arner,
                 {J}oshua {D}. and {Y}ager, {N}eil and {G}ouillart, {E}mmanuelle and
                 {Y}u, {T}ony and {the scikit-image contributors}},
@@ -162,14 +162,14 @@
                   fionagriffin and
                   bmccandless and
                   Marcus Kinsella and
-                  Snyk bot and
+                  {Snyk bot} and
                   Prete and
                   Philipp A. and
-                  Marcio von Muhlen and
+                  Marcio {von Muhlen} and
                   James Taylor and
                   Isaac Virshup and
                   Gökçen Eraslan and
-                  GenevieveHaliburton and
+                  Genevieve Haliburton and
                   Alex Wolf},
   title        = {chanzuckerberg/cellxgene: Release 0.15.0},
   month        = mar,
