+++ paper/paper.bib
@@ -229,7 +229,7 @@
     author = {Yang, Xiaoyu and Bergenholtz, Seth and Maliskova, Lenka and Pebworth, Mark-Phillip and Kriegstein, Arnold R. and Li, Yun and Shen, Yin},
     journal = {PLOS ONE},
     publisher = {Public Library of Science},
-    title = {SMART-Q: An Integrative Pipeline Quantifying Cell Type-Specific RNA Transcription},
+    title = {{SMART-Q}: An Integrative Pipeline Quantifying Cell Type-Specific RNA Transcription},
     year = {2020},
     month = {04},
     volume = {15},
