### 2021-03-24 #2440 @csoneson
👋 @neuromusic - I just sent a [PR](https://github.com/spacetx/starfish/pull/1942) with a couple of small fixes to the manuscript, mainly for the references formatting. Please check and merge if you agree. In addition, JOSS papers are currently [expected](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) to include an explicit `Statement of Need` section that illustrates the research purpose of the software. I think you could easily split off a part of your current `Summary` section to achieve this.

Then, the next steps would be as follows:
- [ ] Check that the Zenodo archive has the correct metadata. In particular, the title and author lists should match those of the paper.

I can then move on to accept the submission. Thanks!
