### 2021-04-05 #2440 @csoneson
👋 @neuromusic - I just wanted to check in to see how things are going, and make sure that the next steps are clear. Once the things in the post above have been done, we can publish this paper. Let me know if you have any questions. Thanks!
