+++ starfish/REQUIREMENTS-STRICT.txt
@@ -88,7 +88,7 @@ tqdm==4.56.2
 trackpy==0.4.2
 traitlets==4.3.3
 typing-extensions==3.7.4.3
-urllib3==1.26.3
+urllib3==1.26.4
 validators==0.18.2
 wcwidth==0.2.5
 webencodings==0.5.1
