+++ requirements/REQUIREMENTS-CI.txt
@@ -137,7 +137,7 @@ traitlets==4.3.3
 twine==3.3.0
 typed-ast==1.4.2
 typing-extensions==3.7.4.3
-urllib3==1.26.3
+urllib3==1.26.4
 validators==0.18.2
 wcwidth==0.2.5
 webencodings==0.5.1
