+++ requirements/REQUIREMENTS-NAPARI-CI.txt
@@ -8,8 +8,8 @@ attrs==20.3.0
 Babel==2.9.0
 backcall==0.2.0
 bleach==3.3.0
-boto3==1.17.10
-botocore==1.20.10
+boto3==1.17.52
+botocore==1.20.52
 cached-property==1.5.2
 cachey==0.2.1
 certifi==2020.12.5
@@ -17,10 +17,10 @@ cffi==1.14.5
 chardet==4.0.0
 click==7.1.2
 cycler==0.10.0
-dask==2021.2.0
+dask==2021.3.0
 dataclasses==0.6
 decorator==4.4.2
-defusedxml==0.6.0
+defusedxml==0.7.1
 diskcache==5.2.1
 docutils==0.16
 entrypoints==0.3
@@ -30,9 +30,9 @@ HeapDict==1.0.1
 idna==2.10
 imageio==2.9.0
 imagesize==1.2.0
-importlib-metadata==3.4.0
+importlib-metadata==3.10.1
 iniconfig==1.1.1
-ipykernel==5.4.3
+ipykernel==5.5.3
 ipython==7.16.1
 ipython-genutils==0.2.0
 ipywidgets==7.6.3
@@ -41,7 +41,7 @@ Jinja2==2.11.3
 jmespath==0.10.0
 joblib==1.0.1
 jsonschema==3.2.0
-jupyter-client==6.1.11
+jupyter-client==6.1.12
 jupyter-core==4.7.1
 jupyterlab-pygments==0.1.2
 jupyterlab-widgets==1.0.0
@@ -53,47 +53,47 @@ mpmath==1.2.1
 napari==0.3.8
 napari-plugin-engine==0.1.9
 napari-svg==0.1.4
-nbclient==0.5.2
+nbclient==0.5.3
 nbconvert==6.0.7
-nbformat==5.1.2
+nbformat==5.1.3
 nest-asyncio==1.5.1
-networkx==2.5
-notebook==6.2.0
+networkx==2.5.1
+notebook==6.3.0
 numpy==1.19.5
 numpydoc==1.1.0
 packaging==20.9
 pandas==1.1.5
 pandocfilters==1.4.3
-parso==0.8.1
+parso==0.8.2
 pexpect==4.8.0
 pickleshare==0.7.5
-Pillow==8.1.1
+Pillow==8.2.0
 pluggy==0.13.1
-prometheus-client==0.9.0
-prompt-toolkit==3.0.16
+prometheus-client==0.10.1
+prompt-toolkit==3.0.18
 psutil==5.8.0
 ptyprocess==0.7.0
 py==1.10.0
 pycparser==2.20
-Pygments==2.8.0
+Pygments==2.8.1
 PyOpenGL==3.1.5
 pyparsing==2.4.7
 PyQt5==5.14.2
 PyQt5-sip==12.8.1
 pyrsistent==0.17.3
-pytest==6.2.2
+pytest==6.2.3
 pytest-qt==3.3.0
-python-dateutil==2.8.0
+python-dateutil==2.8.1
 pytz==2021.1
 PyWavelets==1.1.1
 PyYAML==5.4.1
 pyzmq==22.0.3
-qtconsole==5.0.2
+qtconsole==5.0.3
 QtPy==1.9.0
 read-roi==1.6.0
 regional==1.1.2
 requests==2.25.1
-s3transfer==0.3.4
+s3transfer==0.3.7
 scikit-image==0.15.0
 scikit-learn==0.24.1
 scipy==1.5.4
@@ -104,7 +104,7 @@ showit==1.1.4
 six==1.15.0
 slicedimage==4.1.1
 snowballstemmer==2.1.0
[…]
