+++ requirements/REQUIREMENTS-CI.txt
@@ -8,8 +8,8 @@ attrs==20.3.0
 Babel==2.9.0
 backcall==0.2.0
 bleach==3.3.0
-boto3==1.17.10
-botocore==1.20.10
+boto3==1.17.52
+botocore==1.20.52
 cached-property==1.5.2
 certifi==2020.12.5
 cffi==1.14.5
@@ -17,24 +17,24 @@ chardet==4.0.0
 click==7.1.2
 colorama==0.4.4
 commonmark==0.9.1
-coverage==5.4
+coverage==5.5
 cycler==0.10.0
 dataclasses==0.6
 decorator==4.4.2
-defusedxml==0.6.0
+defusedxml==0.7.1
 diskcache==5.2.1
 docutils==0.16
 entrypoints==0.3
 execnet==1.8.0
-flake8==3.8.4
+flake8==3.9.0
 flake8-import-order==0.18.1
 h5py==3.1.0
 idna==2.10
 imageio==2.9.0
 imagesize==1.2.0
-importlib-metadata==3.4.0
+importlib-metadata==3.10.1
 iniconfig==1.1.1
-ipykernel==5.4.3
+ipykernel==5.5.3
 ipython==7.16.1
 ipython-genutils==0.2.0
 ipywidgets==7.6.3
@@ -43,11 +43,11 @@ Jinja2==2.11.3
 jmespath==0.10.0
 joblib==1.0.1
 jsonschema==3.2.0
-jupyter-client==6.1.11
+jupyter-client==6.1.12
 jupyter-core==4.7.1
 jupyterlab-pygments==0.1.2
 jupyterlab-widgets==1.0.0
-keyring==22.0.1
+keyring==23.0.1
 kiwisolver==1.3.1
 m2r2==0.2.7
 MarkupSafe==1.1.1
@@ -55,53 +55,53 @@ matplotlib==3.3.4
 mccabe==0.6.1
 mistune==0.8.4
 mpmath==1.2.1
-mypy==0.800
+mypy==0.812
 mypy-extensions==0.4.3
-nbclient==0.5.2
+nbclient==0.5.3
 nbconvert==6.0.7
 nbencdec==0.0.10
-nbformat==5.1.2
+nbformat==5.1.3
 nest-asyncio==1.5.1
-networkx==2.5
-notebook==6.2.0
+networkx==2.5.1
+notebook==6.3.0
 numpy==1.19.5
 numpydoc==1.1.0
 packaging==20.9
 pandas==1.1.5
 pandocfilters==1.4.3
-parso==0.8.1
+parso==0.8.2
 pexpect==4.8.0
 pickleshare==0.7.5
-Pillow==8.1.1
+Pillow==8.2.0
 pkginfo==1.7.0
 pluggy==0.13.1
-prometheus-client==0.9.0
-prompt-toolkit==3.0.16
+prometheus-client==0.10.1
+prompt-toolkit==3.0.18
 ptyprocess==0.7.0
 py==1.10.0
-pycodestyle==2.6.0
+pycodestyle==2.7.0
 pycparser==2.20
-pyflakes==2.2.0
-Pygments==2.8.0
+pyflakes==2.3.1
+Pygments==2.8.1
 pyparsing==2.4.7
 pyrsistent==0.17.3
-pytest==6.2.2
+pytest==6.2.3
 pytest-cov==2.11.1
 pytest-forked==1.3.0
 pytest-xdist==2.2.1
-python-dateutil==2.8.0
+python-dateutil==2.8.1
 pytz==2021.1
 PyWavelets==1.1.1
 PyYAML==5.4.1
 pyzmq==22.0.3
 read-roi==1.6.0
-readme-renderer==28.0
[…]
