+++ starfish/REQUIREMENTS-STRICT.txt
@@ -5,8 +5,8 @@ async-generator==1.10
 attrs==20.3.0
 backcall==0.2.0
 bleach==3.3.0
-boto3==1.17.10
-botocore==1.20.10
+boto3==1.17.52
+botocore==1.20.52
 cached-property==1.5.2
 certifi==2020.12.5
 cffi==1.14.5
@@ -15,14 +15,14 @@ click==7.1.2
 cycler==0.10.0
 dataclasses==0.6
 decorator==4.4.2
-defusedxml==0.6.0
+defusedxml==0.7.1
 diskcache==5.2.1
 entrypoints==0.3
 h5py==3.1.0
 idna==2.10
 imageio==2.9.0
-importlib-metadata==3.4.0
-ipykernel==5.4.3
+importlib-metadata==3.10.1
+ipykernel==5.5.3
 ipython==7.16.1
 ipython-genutils==0.2.0
 ipywidgets==7.6.3
@@ -31,7 +31,7 @@ Jinja2==2.11.3
 jmespath==0.10.0
 joblib==1.0.1
 jsonschema==3.2.0
-jupyter-client==6.1.11
+jupyter-client==6.1.12
 jupyter-core==4.7.1
 jupyterlab-pygments==0.1.2
 jupyterlab-widgets==1.0.0
@@ -40,28 +40,28 @@ MarkupSafe==1.1.1
 matplotlib==3.3.4
 mistune==0.8.4
 mpmath==1.2.1
-nbclient==0.5.2
+nbclient==0.5.3
 nbconvert==6.0.7
-nbformat==5.1.2
+nbformat==5.1.3
 nest-asyncio==1.5.1
-networkx==2.5
-notebook==6.2.0
+networkx==2.5.1
+notebook==6.3.0
 numpy==1.19.5
 packaging==20.9
 pandas==1.1.5
 pandocfilters==1.4.3
-parso==0.8.1
+parso==0.8.2
 pexpect==4.8.0
 pickleshare==0.7.5
-Pillow==8.1.1
-prometheus-client==0.9.0
-prompt-toolkit==3.0.16
+Pillow==8.2.0
+prometheus-client==0.10.1
+prompt-toolkit==3.0.18
 ptyprocess==0.7.0
 pycparser==2.20
-Pygments==2.8.0
+Pygments==2.8.1
 pyparsing==2.4.7
 pyrsistent==0.17.3
-python-dateutil==2.8.0
+python-dateutil==2.8.1
 pytz==2021.1
 PyWavelets==1.1.1
 PyYAML==5.4.1
@@ -69,7 +69,7 @@ pyzmq==22.0.3
 read-roi==1.6.0
 regional==1.1.2
 requests==2.25.1
-s3transfer==0.3.4
+s3transfer==0.3.7
 scikit-image==0.15.0
 scikit-learn==0.24.1
 scipy==1.5.4
@@ -80,12 +80,12 @@ showit==1.1.4
 six==1.15.0
 slicedimage==4.1.1
 sympy==1.5.1
-terminado==0.9.2
+terminado==0.9.4
 testpath==0.4.4
 threadpoolctl==2.1.0
 tornado==6.1
-tqdm==4.56.2
-trackpy==0.4.2
+tqdm==4.60.0
+trackpy==0.5.0
 traitlets==4.3.3
 typing-extensions==3.7.4.3
 urllib3==1.26.4
@@ -94,4 +94,4 @@ wcwidth==0.2.5
 webencodings==0.5.1
 widgetsnbextension==3.5.1
 xarray==0.16.2
-zipp==3.4.0
+zipp==3.4.1
