### 2021-04-16 215c340
scikit-image version fix
