+++ paper/paper.md
@@ -61,6 +61,8 @@ These technologies promise to permit biologists to ask new questions about the s
 However, these assays generate very large microscopy datasets which are challenging to process using general microscopy analysis tools.
 Furthermore, many of these assays require specialized analysis to decode gene expression from multiplexed experimental designs.
 
+# Statement of Need
+
 *starfish* is a Python library for processing images generated by microscopy-based spatial transcriptomics assays.
 It lets biologists build scalable pipelines that localize and quantify RNA transcripts in image data generated by any hybridization- or sequencing-based *in situ* transcriptomics method, from classic RNA single-molecule FISH to combinatorial barcoded assays.
 Image processing of an experiment is divided into fields of view (FOV) that correspond to the data produced by a microscope at a single location on a microscope slide.
