### 2021-04-25 #2440 @csoneson
@neuromusic - could you update us on how things are going here? Please note that the archived version of the software should correspond closely to what the reviewers had a chance to look at - in other words, we don't expect there to be lots of new developments going on between the reviewers signing off and the software archive that will be linked with the JOSS publication being generated.
