### 2021-04-30 510241c
moves IntensityTable import
