+++ setup.cfg
@@ -12,8 +12,8 @@ summary = Pipelines and pipeline components for the analysis of image-based tran
 author = Deep Ganguli
 author_email = dganguli@chanzuckerberg.com
 license = MIT
-description-file = README.rst
-description-content-type = text/x-rst; charset=UTF-8
+long_description = README.rst
+long_description_content_type = text/x-rst; charset=UTF-8
 home-page = https://spacetx-starfish.readthedocs.io/en/latest/
 project_urls =
     Bug Tracker = https://github.com/spacetx/starfish/issues
