+++ starfish/core/config/test/test_config.py
@@ -156,11 +156,11 @@ def test_cache_merfish(tmpdir, name, expected, config, monkeypatch):
         setup_config(config, tmpdir, mc)
 
         # Run 1
-        data.MERFISH(use_test_data=True).fov().get_image("primary")
+        data.MERFISH(use_test_data=True).fov().get_image("primary").xarray
 
         # Run 2
         if cache_enabled:
-            data.MERFISH(use_test_data=True).fov().get_image("primary")
+            data.MERFISH(use_test_data=True).fov().get_image("primary").xarray
 
         # Check constraints
         if cache_enabled:
