### 2021-05-03 #2440 @neuromusic
ok, finally cut the release & updated zenodo info.

there was no additional development besides fixing some dependencies for the release.

DOI here: 10.5281/zenodo.4734993
